requirements.txt
fluent-logger==0.9.6
pyyaml==5.4.1